Fixed another bug in MarkdownParse
Previously, if there was a period between the parenthesis, it would be counted as a link, even if it wasn't actually a link.

SO, I began checking for spaces between the parentheses. If there is a space, it would not be counted as a link. So the loop would continue to the next iteration, eventually hitting every space until it found all the links.